LIBWEB-5393. Refactored code based on PHP Code Sniffer recommendations
Renamed "DrupalGateway" interface class to "DrupalGatewayInterface",
and "DrupalGatewayImpl" to "DrupalGateway".

Add PHPDoc blocks, and renames some variables/methods to have better
names and to match PHP capitalization conventions.

https://issues.umd.edu/browse/LIBWEB-5393
