+++ web/modules/custom/umd_staff_directory_rest/umd_staff_directory_rest.module
@@ -4,5 +4,3 @@
  * @file
  * Module file for UMD Staff Directory REST module.
  */
-
-// Intentionally empty.
