+++ web/modules/custom/umd_staff_directory_rest/umd_staff_directory_rest.services.yml
@@ -1,3 +1,3 @@
 services:
   umd_staff_directory_rest.gateway:
-    class: Drupal\umd_staff_directory_rest\impl\DrupalGatewayImpl
+    class: Drupal\umd_staff_directory_rest\impl\DrupalGateway
